@@ -22,6 +22,7 @@
     "date-fns-tz": "^3.2.0",
     "distinct-colors": "^3.0.0",
     "fancy-ansi": "^0.1.3",
+    "graphql": "^16.11.0",
     "graphql-ws": "^5.16.2",
     "lucide-react": "^0.516.0",
     "numeral": "^2.0.6",
@@ -41,38 +42,41 @@
   "devDependencies": {
     "@apollo/react-testing": "^4.0.0",
     "@babel/core": "^7.27.4",
-    "@eslint/js": "^9.28.0",
+    "@eslint/compat": "^1.3.0",
+    "@eslint/js": "^9.29.0",
     "@graphql-codegen/cli": "^5.0.7",
     "@graphql-codegen/client-preset": "^4.8.2",
     "@graphql-codegen/fragment-matcher": "^5.1.0",
     "@graphql-typed-document-node/core": "^3.2.0",
+    "@stylistic/eslint-plugin": "^3.1.0",
     "@testing-library/jest-dom": "^6.6.3",
     "@testing-library/react": "^16.3.0",
-    "@types/node": "^22.15.30",
+    "@types/node": "^24.0.3",
     "@types/numeral": "^2.0.5",
     "@types/react": "^18.3.23",
     "@types/react-dom": "^18.3.7",
     "@types/react-timeago": "^4.1.7",
     "@types/react-window": "^1.8.8",
     "@types/react-window-infinite-loader": "^1.0.9",
-    "@typescript-eslint/eslint-plugin": "^7.18.0",
-    "@typescript-eslint/parser": "^7.18.0",
     "@vitejs/plugin-react-swc": "^3.10.1",
     "autoprefixer": "^10.4.21",
-    "eslint": "^8.57.1",
-    "eslint-config-airbnb": "^19.0.4",
-    "eslint-config-airbnb-typescript": "^18.0.0",
-    "eslint-plugin-import": "^2.31.0",
+    "eslint": "^9.29.0",
+    "eslint-config-airbnb-extended": "^2.1.0",
+    "eslint-config-prettier": "^10.1.5",
+    "eslint-import-resolver-typescript": "^4.4.3",
+    "eslint-plugin-import-x": "^4.15.2",
     "eslint-plugin-jsx-a11y": "^6.10.2",
-    "eslint-plugin-react-hooks": "^4.6.2",
+    "eslint-plugin-prettier": "^5.5.0",
+    "eslint-plugin-react": "^7.37.5",
+    "eslint-plugin-react-hooks": "^5.2.0",
     "eslint-plugin-react-refresh": "^0.4.20",
     "globals": "^15.15.0",
     "history": "^5.3.0",
     "jsdom": "^26.1.0",
     "postcss": "^8.5.4",
     "tailwindcss": "^3.4.17",
-    "typescript": "5.5.4",
-    "typescript-eslint": "^8.33.1",
+    "typescript": "^5.8.3",
+    "typescript-eslint": "^8.34.1",
     "unplugin-fonts": "^1.3.1",
     "vite": "^5.4.19",
     "vite-plugin-svgr": "^4.3.0",
@@ -83,7 +87,8 @@
   "pnpm": {
     "onlyBuiltDependencies": [
       "@swc/core",
-      "esbuild"
+      "esbuild",
+      "unrs-resolver"
     ]
   }
 }
