@@ -17,8 +17,6 @@ import { createContext, useContext, useEffect, useMemo, useState } from 'react';
 import appConfig from '@/app-config';
 import { getBasename, joinPaths } from '@/lib/util';
 
-interface SessionProviderProps extends React.PropsWithChildren { }
-
 const bcName = 'auth/session';
 const bcIn = new BroadcastChannel(bcName);
 const bcOut = new BroadcastChannel(bcName);
@@ -71,7 +69,7 @@ export function useSession() {
  * Session provider (Desktop)
  */
 
-export const SessionProviderDesktop = ({ children }: SessionProviderProps) => {
+export const SessionProviderDesktop = ({ children }: React.PropsWithChildren) => {
   const context = useMemo(() => ({
     session: {
       auth_mode: 'auto',
@@ -92,7 +90,7 @@ export const SessionProviderDesktop = ({ children }: SessionProviderProps) => {
  * Session provider (Cluster)
  */
 
-export const SessionProviderCluster = ({ children }: SessionProviderProps) => {
+export const SessionProviderCluster = ({ children }: React.PropsWithChildren) => {
   const [session, setSession] = useState<Session | undefined>(undefined);
 
   // subscribe to broadcast messages
@@ -137,7 +135,7 @@ export const SessionProviderCluster = ({ children }: SessionProviderProps) => {
  * Session provider
  */
 
-export const SessionProvider = ({ children }: SessionProviderProps) => {
+export const SessionProvider = ({ children }: React.PropsWithChildren) => {
   if (appConfig.environment === 'desktop') {
     return <SessionProviderDesktop>{children}</SessionProviderDesktop>;
   }
